Those changes are small but as far as I can tell they are correct.

+++ include/freerdp/freerdp.h
@@ -75,7 +75,7 @@ struct rdp_inst
 		void ** write_fds, int * write_count);
 	int (* rdp_check_fds)(rdpInst * inst);
 	int (* rdp_send_input_scancode)(rdpInst * inst, RD_BOOL up, RD_BOOL extended, uint8 keyCode);
-	int (* rdp_send_input_unicode)(rdpInst * inst, uint8 character);
+	int (* rdp_send_input_unicode)(rdpInst * inst, uint16 character);
 	int (* rdp_send_input_mouse)(rdpInst * inst, uint16 pointerFlags, uint16 xPos, uint16 yPos);
 	int (* rdp_sync_input)(rdpInst * inst, int toggle_flags);
 	int (* rdp_channel_data)(rdpInst * inst, int chan_id, char * data, int data_size);
